anti raid hehe

+++ cogs/moderation/automod.py
@@ -2,6 +2,10 @@ import discord
 from discord.ext import commands
 from motor.motor_asyncio import AsyncIOMotorClient
 from main import config_var
+import string
+import random
+import discapty
+import asyncio
 
 cluster = AsyncIOMotorClient(config_var['mango_link'])
 cursor = cluster["moderation"]['automod']
@@ -67,5 +71,44 @@ class AutoMod(commands.Cog):
                     return
                 else:
                     if len(message.mentions) > 3:
-                        await message.delete()
-                        await message.channel.send("There's a mass ping. Do something mods")
+                        if check['anti mass mention'] == "on":
+                            await message.delete()
+                            await message.channel.send("There's a mass ping. Do something mods")
+
+    @commands.Cog.listener()
+    async def on_member_join(self, member):
+        check = await cursor.find_one({"guild": member.guild.id})
+        if check is None:
+            return
+        else:
+            if check['anti raid'] == "on":
+                if not member.bot:
+                    embed = discord.Embed(title=f"Hello {member}",
+                                          description="This server has anti raid enabled! If you don't pass this test, you'll get kicked! \n**NOTE THIS IS CASE SENSITIVE**",
+                                          color=discord.Color.red())
+                    embed.set_footer(text="Complete this in 30 seconds")
+
+                    characters = list(string.ascii_letters + string.digits + "!@#$%^&*()")
+                    random.shuffle(characters)
+
+                    password = []
+                    for i in range(10):
+                        password.append(random.choice(characters))
+                    random.shuffle(password)
+
+                    c = "".join(password)
+                    captcha = discapty.Captcha(c)
+                    captcha_image = discord.File(captcha.generate_captcha(), filename="captcha.png")
+                    await member.send(embed=embed, file=captcha_image)
+                    try:
+                        message = await self.bot.wait_for('message', timeout=30.0)
+                    except asyncio.TimeoutError:
+                        await member.send("You FAILED the test. Now you get kicked for being a raider")
+                        await member.kick(reason="Potential Raider")
+                    else:
+                        if message.content == c:
+                            await member.send("You have past the test. Now you're in")
+                            await member.kick(reason="Potential Raider")
+                        else:
+                            await member.send("You FAILED the test. Now you get kicked for being a raider")
+                            await member.kick(reason="Potential Raider")
